requirements.txt
flask
flask-cors
openai
airtable-python-wrapper
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pytest
gunicorn
